requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-WTF==1.1.1
mysql-connector-python==8.1.0
mysqlclient==2.2.0
Werkzeug==2.3.7
email-validator==2.0.0
python-dotenv==1.0.0
Pillow==10.0.0
bcrypt==4.0.1 